// webclient/src/gen/message.ts
export interface PortForward {
  host: string;
  port: number;
}

export interface FileTransfer {
  dir: string;
  show_hidden: boolean;
}

export interface ViewCamera {
}

export interface OSLogin {
  username: string;
  password: string;
}

export interface LoginRequest {
  username: string;
  password: Uint8Array;
  my_id: string;
  my_name: string;
  option?: OptionMessage | undefined;
  union?:
    | { $case: "file_transfer"; file_transfer: FileTransfer }
    | { $case: "port_forward"; port_forward: PortForward }
    | { $case: "view_camera"; view_camera: ViewCamera }
    | { $case: "terminal"; terminal: Terminal }
    | undefined;
  video_ack_required: boolean;
  session_id: bigint;
  version: string;
  os_login?: OSLogin | undefined;
  my_platform: string;
  hwid: Uint8Array;
  avatar: string;
}

export interface Terminal {
  /** Service ID for reconnecting to existing session */
  service_id: string;
}
// ...
export interface SupportedDecoding {
  ability_vp9: number;
  ability_h264: number;
  ability_h265: number;
  prefer: SupportedDecoding_PreferCodec;
  ability_vp8: number;
  ability_av1: number;
  i444?: CodecAbility | undefined;
  prefer_chroma: Chroma;
}

export enum SupportedDecoding_PreferCodec {
  Auto = 0,
  VP9 = 1,
  H264 = 2,
  H265 = 3,
  VP8 = 4,
  AV1 = 5,
}

export interface OptionMessage {
  image_quality: ImageQuality;
  lock_after_session_end: OptionMessage_BoolOption;
  show_remote_cursor: OptionMessage_BoolOption;
  privacy_mode: OptionMessage_BoolOption;
  block_input: OptionMessage_BoolOption;
  custom_image_quality: number;
  disable_audio: OptionMessage_BoolOption;
  disable_clipboard: OptionMessage_BoolOption;
  enable_file_transfer: OptionMessage_BoolOption;
  supported_decoding?: SupportedDecoding | undefined;
  custom_fps: number;
  disable_keyboard: OptionMessage_BoolOption;
  /**
   * Position 13 is used for Resolution. Remove later.
   * Resolution custom_resolution = 13;
   * BoolOption support_windows_specific_session = 14;
   * starting from 15 please, do not use removed fields
   */
  follow_remote_cursor: OptionMessage_BoolOption;
  follow_remote_window: OptionMessage_BoolOption;
  disable_camera: OptionMessage_BoolOption;
  terminal_persistent: OptionMessage_BoolOption;
  show_my_cursor: OptionMessage_BoolOption;
}

export enum OptionMessage_BoolOption {
  NotSet = 0,
  No = 1,
  Yes = 2,
}

export interface TestDelay {
  time: bigint;
  from_client: boolean;
  last_delay: number;
  target_bitrate: number;
}

export interface PublicKey {
  asymmetric_value: Uint8Array;
  symmetric_value: Uint8Array;
}

export interface SignedId {
  id: Uint8Array;
}
// ...
export interface Message {
  union?:
    | { $case: "signed_id"; signed_id: SignedId }
    | { $case: "public_key"; public_key: PublicKey }
    | { $case: "test_delay"; test_delay: TestDelay }
    | { $case: "video_frame"; video_frame: VideoFrame }
    | { $case: "login_request"; login_request: LoginRequest }
    | { $case: "login_response"; login_response: LoginResponse }
    | { $case: "hash"; hash: Hash }
    | { $case: "mouse_event"; mouse_event: MouseEvent }
    | { $case: "audio_frame"; audio_frame: AudioFrame }
    | { $case: "cursor_data"; cursor_data: CursorData }
    | { $case: "cursor_position"; cursor_position: CursorPosition }
    | { $case: "cursor_id"; cursor_id: bigint }
    | { $case: "key_event"; key_event: KeyEvent }
    | { $case: "clipboard"; clipboard: Clipboard }
    | { $case: "file_action"; file_action: FileAction }
    | { $case: "file_response"; file_response: FileResponse }
    | { $case: "misc"; misc: Misc }
    | { $case: "cliprdr"; cliprdr: Cliprdr }
    | { $case: "message_box"; message_box: MessageBox }
    | { $case: "switch_sides_response"; switch_sides_response: SwitchSidesResponse }
    | { $case: "voice_call_request"; voice_call_request: VoiceCallRequest }
    | { $case: "voice_call_response"; voice_call_response: VoiceCallResponse }
    | { $case: "peer_info"; peer_info: PeerInfo }
    | { $case: "pointer_device_event"; pointer_device_event: PointerDeviceEvent }
    | { $case: "auth_2fa"; auth_2fa: Auth2FA }
    | { $case: "multi_clipboards"; multi_clipboards: MultiClipboards }
    | { $case: "screenshot_request"; screenshot_request: ScreenshotRequest }
    | { $case: "screenshot_response"; screenshot_response: ScreenshotResponse }
    | { $case: "terminal_action"; terminal_action: TerminalAction }
    | { $case: "terminal_response"; terminal_response: TerminalResponse }
    | undefined;
}
// ...
export const PortForward: MessageFns<PortForward> = {
  encode(message: PortForward, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.host !== "") {
      writer.uint32(10).string(message.host);
    }
    if (message.port !== 0) {
      writer.uint32(16).int32(message.port);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PortForward {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePortForward();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.host = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.port = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<PortForward>, I>>(base?: I): PortForward {
    return PortForward.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PortForward>, I>>(object: I): PortForward {
    const message = createBasePortForward();
    message.host = object.host ?? "";
    message.port = object.port ?? 0;
    return message;
  },
};

function createBaseFileTransfer(): FileTransfer {
  return { dir: "", show_hidden: false };
}

export const FileTransfer: MessageFns<FileTransfer> = {
  encode(message: FileTransfer, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.dir !== "") {
      writer.uint32(10).string(message.dir);
    }
    if (message.show_hidden !== false) {
      writer.uint32(16).bool(message.show_hidden);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileTransfer {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileTransfer();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.dir = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.show_hidden = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<FileTransfer>, I>>(base?: I): FileTransfer {
    return FileTransfer.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileTransfer>, I>>(object: I): FileTransfer {
    const message = createBaseFileTransfer();
    message.dir = object.dir ?? "";
    message.show_hidden = object.show_hidden ?? false;
    return message;
  },
};

function createBaseViewCamera(): ViewCamera {
  return {};
}

export const ViewCamera: MessageFns<ViewCamera> = {
  encode(_: ViewCamera, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ViewCamera {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseViewCamera();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ViewCamera>, I>>(base?: I): ViewCamera {
    return ViewCamera.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ViewCamera>, I>>(_: I): ViewCamera {
    const message = createBaseViewCamera();
    return message;
  },
};
// ...
export const SupportedDecoding: MessageFns<SupportedDecoding> = {
  encode(message: SupportedDecoding, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ability_vp9 !== 0) {
      writer.uint32(8).int32(message.ability_vp9);
    }
    if (message.ability_h264 !== 0) {
      writer.uint32(16).int32(message.ability_h264);
    }
    if (message.ability_h265 !== 0) {
      writer.uint32(24).int32(message.ability_h265);
    }
    if (message.prefer !== 0) {
      writer.uint32(32).int32(message.prefer);
    }
    if (message.ability_vp8 !== 0) {
      writer.uint32(40).int32(message.ability_vp8);
    }
    if (message.ability_av1 !== 0) {
      writer.uint32(48).int32(message.ability_av1);
    }
    if (message.i444 !== undefined) {
      CodecAbility.encode(message.i444, writer.uint32(58).fork()).join();
    }
    if (message.prefer_chroma !== 0) {
      writer.uint32(64).int32(message.prefer_chroma);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SupportedDecoding {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSupportedDecoding();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.ability_vp9 = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.ability_h264 = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.ability_h265 = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.prefer = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.ability_vp8 = reader.int32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.ability_av1 = reader.int32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.i444 = CodecAbility.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.prefer_chroma = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<SupportedDecoding>, I>>(base?: I): SupportedDecoding {
    return SupportedDecoding.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SupportedDecoding>, I>>(object: I): SupportedDecoding {
    const message = createBaseSupportedDecoding();
    message.ability_vp9 = object.ability_vp9 ?? 0;
    message.ability_h264 = object.ability_h264 ?? 0;
    message.ability_h265 = object.ability_h265 ?? 0;
    message.prefer = object.prefer ?? 0;
    message.ability_vp8 = object.ability_vp8 ?? 0;
    message.ability_av1 = object.ability_av1 ?? 0;
    message.i444 = (object.i444 !== undefined && object.i444 !== null)
      ? CodecAbility.fromPartial(object.i444)
      : undefined;
    message.prefer_chroma = object.prefer_chroma ?? 0;
    return message;
  },
};